Created Add Individual Modal.
