+++ src/app/main-dashboard/add-person/add-person-content.html
@@ -1,0 +1,24 @@
+<mat-dialog-content>
+    <mat-card class="dialog-card">
+        <div align="end">
+            <button class = "closebtn" matIconButton mat-dialog-close aria-label="Close dialog" align="end">
+                <mat-icon>close</mat-icon><span class="spacer"></span>
+            </button>
+        </div>
+        <mat-card-header>
+            <mat-card-title>Add Person</mat-card-title>
+        </mat-card-header>
+        <mat-card-content>
+            <p>Choose a person type: </p>
+            <div>
+                <mat-dialog-actions>
+                    <button mat-raised-button style="margin: 10px;" (click)="personTypeClick('Individual')">Individual</button>
+                    <button mat-raised-button style="margin: 10px;">Organization</button>
+                </mat-dialog-actions>
+            </div>
+            <div><br></div>
+        </mat-card-content> 
+    </mat-card>
+    <br>
+</mat-dialog-content>
+
