+++ src/app/main-dashboard/add-person/individual-content.html
@@ -6,18 +6,16 @@
             </button>
         </div>
         <mat-card-header>
-            <mat-card-title>Add Person</mat-card-title>
+            <mat-card-title> Add Individual</mat-card-title>
         </mat-card-header>
         <mat-card-content>
-            <p>Choose a person type: </p>
+            <p>This is where the user will enter contact information and disability info about a given individual</p>
+            
             <div>
-                <button mat-button>Individual</button>
-                <button mat-button>Organization</button>
+                <mat-dialog-actions>
+                </mat-dialog-actions>
             </div>
             <div><br></div>
-        <mat-dialog-actions align="end">
-            <button mat-raised-button color="primary">Confirm</button>
-        </mat-dialog-actions>
         </mat-card-content> 
     </mat-card>
     <br>
